requirements.txt
asgiref==3.2.7
boto3==1.12.46
botocore==1.15.46
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-classy-tags==1.0.0
django-cms==3.7.3
django-cors-headers==3.4.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-formtools==2.2
django-heroku==0.3.1
django-sekizai==1.1.0
django-snippet==1.0.1
django-storages==1.9.1
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-snippet==2.3.0
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
jmespath==0.9.5
Markdown==3.2.2
oauthlib==3.1.0
Pillow==7.1.1
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
talk==0.1.0
urllib3==1.25.9
whitenoise==5.0.1
zipp==3.1.0
